refactor : separate class from playgame methos base on SRP
